+++ packages/components/form-field/src/FormField.tsx
@@ -13,15 +13,25 @@ export interface FormFieldProps
    * Change the component to the HTML tag or custom component of the only child. This will merge the original component props with the props of the supplied element/component and change the underlying DOM node.
    */
   asChild?: boolean
+  disabled?: boolean
 }
 
 export const FormField = forwardRef<HTMLDivElement, FormFieldProps>(
-  ({ className, name, state, isRequired = false, asChild = false, ...others }, ref) => {
+  (
+    { className, disabled = false, name, state, isRequired = false, asChild = false, ...others },
+    ref
+  ) => {
     const id = useId()
     const Component = asChild ? Slot : 'div'
 
     return (
-      <FormFieldProvider id={id} name={name} isRequired={isRequired} state={state}>
+      <FormFieldProvider
+        id={id}
+        name={name}
+        isRequired={isRequired}
+        disabled={disabled}
+        state={state}
+      >
         <Component
           ref={ref}
           data-spark-component="form-field"