+++ packages/components/checkbox/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.19.1](https://github.com/adevinta/spark/compare/@spark-ui/checkbox@1.19.0...@spark-ui/checkbox@1.19.1) (2023-07-26)
+
+**Note:** Version bump only for package @spark-ui/checkbox
+
+# [1.19.0](https://github.com/adevinta/spark/compare/@spark-ui/checkbox@1.18.2...@spark-ui/checkbox@1.19.0) (2023-07-26)
+
+### Features
+
+- update form field version ([738b1b3](https://github.com/adevinta/spark/commit/738b1b3f86e541e10aa2499d5840972208370703))
+
+## [1.18.2](https://github.com/adevinta/spark/compare/@spark-ui/checkbox@1.18.1...@spark-ui/checkbox@1.18.2) (2023-07-26)
+
+**Note:** Version bump only for package @spark-ui/checkbox
+
+## [1.18.1](https://github.com/adevinta/spark/compare/@spark-ui/checkbox@1.18.0...@spark-ui/checkbox@1.18.1) (2023-07-24)
+
+### Bug Fixes
+
+- updating deprecated formField versions ([a8aacc8](https://github.com/adevinta/spark/commit/a8aacc87818dcf7c7e5ca1ad75e9cd42bedeb4df))
+
 # [1.18.0](https://github.com/adevinta/spark/compare/@spark-ui/checkbox@1.17.18...@spark-ui/checkbox@1.18.0) (2023-07-21)
 
 ### Features